+++ common/Octets.h
@@ -5,7 +5,8 @@
  * Keep it simple at first 
  *************************************************************/
 
-#include <cstddef>
+#include <stddef.h>
+#include <stdint.h>
 
 namespace catman
 {
@@ -18,32 +19,32 @@ class Octets
 	{
 	public:
 		Rep();
-		size_t capacity() const;
-		size_t size() const;
-		void setSize(size_t newSize);
+		uint32_t capacity() const;
+		uint32_t size() const;
+		void setSize(uint32_t newSize);
 		void incRef();
 		void decRef();
 		void* data();
 		void* unique();
-		void* reserve(size_t size);
-		static Rep* create(size_t capacity);
+		void* reserve(uint32_t size);
+		static Rep* create(uint32_t capacity);
 	private:
 		Rep* clone();
-		static size_t alignSize(size_t size);
-		static void* operator new(size_t size, size_t extra);
+		static uint32_t alignSize(uint32_t size);
+		static void* operator new(size_t size, uint32_t extra);
 		static void operator delete(void *p);
 	public:
 		static Rep null;
 	private:
-		size_t m_capacity;
-		size_t m_size;
-		size_t m_ref;
+		uint32_t m_capacity;
+		uint32_t m_size;
+		uint32_t m_ref;
 		
 	};
 public:
 	Octets();
-	Octets(size_t size);
-	Octets(const void *buff, size_t len);
+	Octets(uint32_t size);
+	Octets(const void *buff, uint32_t len);
 	Octets(const void *beginPos, const void *endPos);
 	Octets(const Octets &other);
 	virtual ~Octets();
@@ -55,16 +56,16 @@ public:
 	void* end();
 	const void* begin() const;
 	const void* end() const;
-	size_t size() const;
-	size_t capacity() const;
+	uint32_t size() const;
+	uint32_t capacity() const;
 	Octets& clear();
-	Octets& erase(size_t pos, size_t len);
+	Octets& erase(uint32_t pos, uint32_t len);
 	Octets& erase(void *beginPos, void *endPos);
-	Octets& insert(void *pos, const void *buff, size_t len);
+	Octets& insert(void *pos, const void *buff, uint32_t len);
 	Octets& insert(void *pos, const void *beginPos, const void *endPos);
-	Octets& reserve(size_t size);
-	Octets& resize(size_t size);
-	Octets& replace(const void *buff, size_t len);
+	Octets& reserve(uint32_t size);
+	Octets& resize(uint32_t size);
+	Octets& replace(const void *buff, uint32_t len);
 private:
 	Rep* rep() const;
 private:
